microhapDB-backend/requirements.txt
fastapi==0.68.2
pydantic==1.10.18
uvicorn==0.15.0
python-multipart==0.0.5
aiofiles==24.1.0
SQLAlchemy==2.0.35
pandas==2.2.2
jinja2==3.0.3
upsetplot==0.9.0
asyncpg==0.29.0
httpx==0.27.2
fastapi-login==1.10.1
python-jose==3.3.0
passlib==1.7.4
alembic==1.13.1
python-dotenv==1.0.0